backend/requirements/base.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0
psycopg[binary]>=3.1.0
django-environ>=0.11.2
channels>=4.2.0
channels-redis>=4.1.0
redis>=5.0.0
celery>=5.3.4
django-celery-beat>=2.5.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
Pillow>=10.1.0
requests>=2.31.0
django-jazzmin>=3.0.0
reportlab>=4.0.0
cryptography>=41.0.0
django-auditlog>=3.0.0
django-storages[s3]>=1.14.0
boto3>=1.34.0

uvicorn[standard]>=0.30.0
prometheus-client>=0.16.0
python-json-logger>=2.0.7

# OpenTelemetry (activé via OTEL_ENABLED=true)
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-django>=0.49b0
opentelemetry-instrumentation-celery>=0.49b0
opentelemetry-instrumentation-requests>=0.49b0
opentelemetry-exporter-otlp>=1.27.0
